send_master_update in join_to_cluster to stop services in the joined server
